ordenando codigo y mejorando
