+++ app/controllers/concerns/language.rb
@@ -1,0 +1,17 @@
+module Language
+  extend ActiveSupport::Concern
+
+  included do
+    around_action :switch_locale
+
+    def switch_locale(&action)
+      I18n.with_locale(locale_from_header, &action)
+    end
+
+    private
+
+    def locale_from_header
+      request.env['HTTP_ACCEPT_LANGUAGE']&.scan(/^[a-z]{2}/)&.first
+    end
+  end
+end
