+++ app/models/product.rb
@@ -19,4 +19,8 @@ class Product < ApplicationRecord
 
   belongs_to :category
   belongs_to :user, default: -> { Current.user }
+
+  def owner?
+    user_id == Current.user.id
+  end
 end
