+++ app/views/products/show.html.erb
@@ -5,6 +5,6 @@
 <%= image_tag @product.photo, width: 400 if @product.photo.attached? %>
 
 
-<%= link_to 'Editar producto', edit_product_path(@product) if @product.user == Current.user %>
+<%= link_to 'Editar producto', edit_product_path(@product) if @product.owner? %>
 
-<%= link_to 'Eliminar producto', product_path(@product), data: { turbo_method: :delete, turbo_confirm: t('common.confirm')} if @product.user == Current.user %>
+<%= link_to 'Eliminar producto', product_path(@product), data: { turbo_method: :delete, turbo_confirm: t('common.confirm')} if @product.owner? %>
